+++ CoreLab/labs/labs/LabEight.java
@@ -24,12 +24,12 @@ public class LabEight {
 		File outputFile2 = new File(WORK_DIR+"w_5-high-pass.png");
 		File outputFile3 = new File(WORK_DIR+"w_5-high-boost.png");
 	
-		double[][] lpFilter = ConvolveLib.KERNEL_LAB7_W5;
-		double[][] apFilter = ConvolveLib.KERNEL_ALLPASS_3;
-		double[][] hpFilter = ConvolveLib.kernelDifference(apFilter, lpFilter);
-		double[][] hbFilter = ConvolveLib.kernelSum(apFilter, 
+		float[][] lpFilter = ConvolveLib.KERNEL_LAB7_W5;
+		float[][] apFilter = ConvolveLib.KERNEL_ALLPASS_3;
+		float[][] hpFilter = ConvolveLib.kernelDifference(apFilter, lpFilter);
+		float[][] hbFilter = ConvolveLib.kernelSum(apFilter, 
 								ConvolveLib.multiplyScalar(
-								ConvolveLib.KERNEL_LAB7_W5, 2.0/9.0));
+								ConvolveLib.KERNEL_LAB7_W5, 2f/9f));
 				
 		// low pass filtering
 		cproc.readWorkingImage(baseFile);
@@ -53,12 +53,12 @@ public class LabEight {
 		File outputFile2 = new File(WORK_DIR+"intensity-high-pass.png");
 		File outputFile3 = new File(WORK_DIR+"intensity-high-boost.png");
 	
-		double[][] lpFilter = ConvolveLib.KERNEL_LAB7_W5;
-		double[][] apFilter = ConvolveLib.KERNEL_ALLPASS_3;
-		double[][] hpFilter = ConvolveLib.kernelDifference(apFilter, lpFilter);
-		double[][] hbFilter = ConvolveLib.kernelSum(apFilter, 
+		float[][] lpFilter = ConvolveLib.KERNEL_LAB7_W5;
+		float[][] apFilter = ConvolveLib.KERNEL_ALLPASS_3;
+		float[][] hpFilter = ConvolveLib.kernelDifference(apFilter, lpFilter);
+		float[][] hbFilter = ConvolveLib.kernelSum(apFilter, 
 								ConvolveLib.multiplyScalar(
-								ConvolveLib.KERNEL_LAB7_W5, 2.0/9.0));
+								ConvolveLib.KERNEL_LAB7_W5, 2f/9f));
 		
 		// low pass filtering
 		cproc.readWorkingImage(baseFile);
@@ -132,9 +132,9 @@ public class LabEight {
 		cproc.writeWorkingImage(outputFile4);
 	}
 	
-	private static void convolveIntensity(BufferedImage img, final double[][] kernel) {
+	private static void convolveIntensity(BufferedImage img, final float[][] kernel) {
 		ImageProcessor colorProc = new ImageProcessor(img); 
-		double[][][] hsi = colorProc.getHsi();
+		float[][][] hsi = colorProc.getHsi();
 		
 		hsi[2] = ImageProcessor.convolveArrayPrimative(kernel, hsi[2]);
